Rollback model creation when properties are invalid (#215)
When creating a model with properties, if creating properties fails the
client swallows the error and logs an error message, leaving a model
with no properties. This is not a great experience, as the user believes
the model was created successfully; however future requests to create
records from the model will fail because there are no properties.

This PR updates the client to delete the newly created mode in the case
that properties cannot be created and raise an exception. This should
make it easier for users to retry without have to manually delete the
model or update properties on an existing model.